redirection hook fix
