+++ src/app/actions.js
@@ -3,7 +3,6 @@ import prisma from "@/lib/prisma";
 import {
   S3Client,
   PutObjectCommand,
-  PutObjectAclCommand,
   DeleteObjectCommand,
 } from "@aws-sdk/client-s3";
 import { getSignedUrl } from "@aws-sdk/s3-request-presigner";
