--- src/lib/nosession.js
@@ -1,15 +1,0 @@
-"use client";
-import { redirect, useRouter } from "next/navigation";
-import { useEffect } from "react";
-
-const useRedirectNoSession = (session) => {
-  const router = useRouter();
-
-  useEffect(() => {
-    if (!session) {
-      redirect("/");
-    }
-  }, [session, router]);
-};
-
-export default useRedirectNoSession;
